refactor: move autocomplete dataset and search
